requirements.txt
torch>=1.9.0
pandas>=1.3.0
numpy>=1.19.5
scikit-learn>=0.24.2
pyyaml>=5.4.1
